==> Preprocessing/3_Extract_Query_Document_Pairs.ipynb <==
import pandas as pd
import re
from bs4 import BeautifulSoup

# Load the dataset
df = pd.read_pickle('dataset_raw.pkl')

df.head()

df.info()

BAM_regex = r'B[ ]?[ÖO][ ]?L[ ]?G[ ]?E A[ ]?D[ ]?L[ ]?[İI][ ]?Y[ ]?E M[ ]?A[ ]?H[ ]?K[ ]?E[ ]?M[ ]?E[ ]?S[ ]?[Iİ]'

count = df[df['extracted_data'].str.contains(BAM_regex)].shape[0]
print(count)

# Discarding "Bölge Adliye Mahkemesi" records, since they they differ in formating
df = df[~df['extracted_data'].str.contains(BAM_regex)]

df.info()

query_regex = r'<br>([^<]*dava dilekçe[^<]*)<br>(.*)'

count = df[df['extracted_data'].str.contains(query_regex)].shape[0]
print(count)

# Discarding rows that dont have our desired query format
df = df[df['extracted_data'].str.contains(query_regex)]

# Extract the query and document pairs
query_doc_df = df['extracted_data'].str.extract(query_regex)
query_doc_df.columns = ['query', 'document']

query_doc_df.head()

query_doc_df.info()

# Remove html tags from the data

query_doc_df = query_doc_df.map(lambda s: BeautifulSoup(s).text)
# query_doc_df = query_doc_df.map(lambda x: re.sub('<.*?>', '', str(x))) # an alternative, might be less robust

# At this point, we've eliminated BAM judgments, extracted query-document pairs and removed the HTML tags. 

query_doc_df.head()

# Let's save our dataset in this state into a pickle
query_doc_df.to_pickle('dataset_qdpairs.pkl')
