==> Preprocessing/4_Extract_BERTurk_Embeddings.ipynb <==
from sentence_transformers import SentenceTransformer
import pandas as pd
import numpy as np
import os
from tqdm.autonotebook import tqdm
tqdm.pandas(desc="progress-bar")

model = SentenceTransformer('emrecan/bert-base-turkish-cased-mean-nli-stsb-tr')

model.max_seq_length = 512

device = 'cuda'
model.to(device)

df = pd.read_pickle("Steps/dataset_qdpairs.pkl")

df.head()

import os

directory = "bert_doc_embedding_chunks"
if not os.path.exists(directory):
    os.makedirs(directory)

chunk_size = 10240
batch_size = 128
num_chunks = len(df) // chunk_size + 1

for chunk in tqdm(range(num_chunks), desc="Processing chunks"):
    start_idx = chunk * chunk_size
    end_idx = min((chunk + 1) * chunk_size, len(df))
    chunk_df = df[start_idx:end_idx]
    
    chunk_embeddings = model.encode(chunk_df.document.to_list(), batch_size=batch_size, normalize_embeddings=True)
    
    # Save embeddings to file
    file_path = f"bert_doc_embedding_chunks/doc_embeddings_curchunk_{chunk}.npy"
    np.save(file_path, chunk_embeddings)
    
    print(f"Processed chunk {chunk+1}/{num_chunks}. Saved embeddings to {file_path}")


directory = "bert_doc_embedding_chunks"

# Initialize an empty list to store the chunk embeddings
chunk_embeddings_list = []

# Iterate over the chunks and load the embeddings
for chunk in tqdm(range(num_chunks), desc="Loading embeddings"):
    file_path = f"bert_doc_embedding_chunks/doc_embeddings_curchunk_{chunk}.npy"
    chunk_embeddings = np.load(file_path)
    chunk_embeddings_list.append(chunk_embeddings)

# Concatenate the chunk embeddings into a single numpy array
doc_embeddings = np.concatenate(chunk_embeddings_list)

# Save the doc embeddings as doc_embeddings.npy
np.save("doc_embeddings.npy", doc_embeddings)

print("Merged and saved the embeddings successfully.")
