==> Preprocessing/5_Preprocess_for_BM25.ipynb <==
import pandas as pd
import gensim as gs
import numpy as np
import sklearn
import rank_bm25
import string
from tqdm.autonotebook import tqdm
tqdm.pandas(desc="progress-bar")
from rank_bm25 import BM25Okapi
import nltk
nltk.download('stopwords')
from nltk.corpus import stopwords
import Stemmer

df = pd.read_pickle("dataset_qdpairs.pkl")

df.head()

def preprocess_str(str_to_process):
    # This is a turkish stemmer, doesn't work perfect but it is consistent at least
    stemmer = Stemmer.Stemmer('turkish')
    
    str_result = str_to_process
    # Remove non-chars
    str_result = gs.parsing.preprocessing.strip_multiple_whitespaces(gs.parsing.preprocessing.strip_numeric(gs.parsing.preprocessing.strip_non_alphanum(str_result)))
    # Lowercase str
    str_result = str_result.lower()
    # Remove stopwords
    str_result = gs.parsing.preprocessing.remove_stopwords(s=str_result, stopwords=stopwords.words("turkish"))
    # Split str
    str_result = str_result.split()
    # Stem words
    str_result = stemmer.stemWords(str_result)

    return str_result

def preprocess_df(df):
    # This is a turkish stemmer, doesn't work perfect but it is consistent at least
    stemmer = Stemmer.Stemmer('turkish')

    # Remove non-chars
    df = df.map(gs.parsing.preprocessing.strip_non_alphanum).map(
                                                gs.parsing.preprocessing.strip_numeric).map(gs.parsing.preprocessing.strip_multiple_whitespaces)
    # Lowercase str
    df = df.map(lambda s: s.lower())
    # Remove stopwords
    df = df.map(lambda s: gs.parsing.preprocessing.remove_stopwords(s=s, stopwords=stopwords.words("turkish")))
    # Split str
    df = df.map(lambda s: s.split())


    # Stem words - it takes long
    df = df.map(stemmer.stemWords)

    return df

import pickle

chunk_size = 20000
num_chunks = len(df) // chunk_size + 1 # should've been +1

for i in range(num_chunks):
    print(f"Chunk {i}")
    start_idx = i * chunk_size
    end_idx = (i + 1) * chunk_size
    chunk = df.iloc[start_idx:end_idx]
    preprocessed_chunk = preprocess_df(chunk)
    preprocessed_chunk.to_pickle(f"chunks/preprocessed_chunk_{i}.pkl")
    del preprocessed_chunk

import pandas as pd
import glob

# Get a list of all the pickles
pickle_files = glob.glob('chunks/preprocessed_chunk_*.pkl')
sorted_pickle_files = sorted(pickle_files, key=lambda x: int(x.split('_')[-1].split('.')[0]))

# Read each pickle and concatenate them into a single dataframe
dfs = []
for pickle_file in sorted_pickle_files:
    print(pickle_file)
    df_chunk = pd.read_pickle(pickle_file)
    dfs.append(df_chunk)

df_merged = pd.concat(dfs)

# Write the merged dataframe back to a pickle file
df_merged.to_pickle('processed_dataset/dataset_qdpairs_processed.pkl')
