==> Preprocessing/6_Extract_BM25_Indexes.ipynb <==
from gensim.corpora import Dictionary
from gensim.models import TfidfModel, OkapiBM25Model
from gensim.similarities import SparseMatrixSimilarity
import numpy as np

import pandas as pd

df = pd.read_pickle("processed_dataset/dataset_qdpairs_processed.pkl")
df.head()

df.info()

corpus = df['document'].to_list()
dictionary = Dictionary(corpus)
bm25_model = OkapiBM25Model(dictionary=dictionary)
bm25_corpus = bm25_model[list(map(dictionary.doc2bow, corpus))]
bm25_index = SparseMatrixSimilarity(bm25_corpus, num_docs=len(corpus), num_terms=len(dictionary), normalize_queries=False, normalize_documents=False)

import os
import pickle

# Create the directory if it doesn't exist
if not os.path.exists('gensim_bm25_pickles'):
    os.makedirs('gensim_bm25_pickles')

# Pickle the objects
pickle.dump(corpus, open('gensim_bm25_pickles/corpus.pkl', 'wb'))
pickle.dump(dictionary, open('gensim_bm25_pickles/dictionary.pkl', 'wb'))
pickle.dump(bm25_model, open('gensim_bm25_pickles/bm25_model.pkl', 'wb'))
pickle.dump(bm25_corpus, open('gensim_bm25_pickles/bm25_corpus.pkl', 'wb'))
pickle.dump(bm25_index, open('gensim_bm25_pickles/bm25_index.pkl', 'wb'))
